+++ console/console-init/ui/src/components/Pagination/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
+export * from "./Pagination";
