+++ console/console-init/ui/src/modules/address-space/AddressSpacePage.tsx
@@ -7,7 +7,6 @@ import * as React from "react";
 import { useLocation, useHistory } from "react-router";
 import { useDocumentTitle, useA11yRouteChange } from "use-patternfly";
 import {
-  Pagination,
   PageSection,
   PageSectionVariants,
   Grid,
@@ -25,6 +24,7 @@ import { DELETE_ADDRESS_SPACE } from "graphql-module/queries";
 import { compareObject } from "utils";
 import { useStoreContext, types, MODAL_TYPES } from "context-state-reducer";
 import { getHeaderForDeleteDialog, getDetailForDeleteDialog } from "./utils";
+import { Pagination } from "components";
 
 export default function AddressSpacePage() {
   const { dispatch } = useStoreContext();
@@ -44,7 +44,6 @@ export default function AddressSpacePage() {
   const [sortDropDownValue, setSortDropdownValue] = React.useState<ISortBy>();
   const [isCreateWizardOpen, setIsCreateWizardOpen] = React.useState(false);
   const location = useLocation();
-  const history = useHistory();
   const searchParams = new URLSearchParams(location.search);
   const page = parseInt(searchParams.get("page") || "", 10) || 1;
   const perPage = parseInt(searchParams.get("perPage") || "", 10) || 10;
@@ -60,47 +59,6 @@ export default function AddressSpacePage() {
       awaitRefetchQueries: true
     }
   );
-
-  const setSearchParam = React.useCallback(
-    (name: string, value: string) => {
-      searchParams.set(name, value.toString());
-    },
-    [searchParams]
-  );
-
-  const handlePageChange = React.useCallback(
-    (_: any, newPage: number) => {
-      setSearchParam("page", newPage.toString());
-      history.push({
-        search: searchParams.toString()
-      });
-    },
-    [setSearchParam, history, searchParams]
-  );
-
-  const handlePerPageChange = React.useCallback(
-    (_: any, newPerPage: number) => {
-      setSearchParam("page", "1");
-      setSearchParam("perPage", newPerPage.toString());
-      history.push({
-        search: searchParams.toString()
-      });
-    },
-    [setSearchParam, history, searchParams]
-  );
-
-  const renderPagination = (page: number, perPage: number) => {
-    return (
-      <Pagination
-        itemCount={totalAddressSpaces}
-        perPage={perPage}
-        page={page}
-        onSetPage={handlePageChange}
-        variant="top"
-        onPerPageSelect={handlePerPageChange}
-      />
-    );
-  };
 
   const deleteAddressSpace = async (
     addressSpace: IAddressSpace,
@@ -219,7 +177,12 @@ export default function AddressSpacePage() {
           />
         </GridItem>
         <GridItem span={5}>
-          {totalAddressSpaces > 0 && renderPagination(page, perPage)}
+          <Pagination
+            itemCount={totalAddressSpaces}
+            variant={"top"}
+            page={page}
+            perPage={perPage}
+          />
         </GridItem>
       </Grid>
       <Divider />
@@ -241,7 +204,12 @@ export default function AddressSpacePage() {
         onSelectAddressSpace={onSelectAddressSpace}
         onSelectAllAddressSpace={onSelectAllAddressSpace}
       />
-      {totalAddressSpaces > 0 && renderPagination(page, perPage)}
+      <Pagination
+        itemCount={totalAddressSpaces}
+        variant={"top"}
+        page={page}
+        perPage={perPage}
+      />
     </PageSection>
   );
 }
